deprecated/requirements.txt
requests
flask
minio
pandas
pystardog
